fix(controllers): use fileId and add validation
refactor(routes): convert routes into RESTful structure

+++ server/src/controllers/fileControllers/fileController.js
@@ -1,5 +1,5 @@
 import File from '../../models/file.model.js'
-import { badRequest, customError, deleted,  success, updated } from '../../utils/response.utils.js'
+import { badRequest, customError, deleted, success, updated } from '../../utils/response.utils.js'
 
 
 const createFileController = async (req, res, next) => {
@@ -24,11 +24,11 @@ const createFileController = async (req, res, next) => {
 
 const readFileController = async (req, res, next) => {
     try {
-        const { fileName } = req.params;
+        const { fileId } = req.params;
 
-        const file = await File.findOne({ name: fileName, user: req.userId });
+        const file = await File.findOne({ _id: fileId, user: req.userId });
 
-        if (!file) return customError(res, "File not found");
+        if (!file) return customError(res, 404, "File not found");
 
         return success(res, file, "File read");
     } catch (err) {
@@ -37,7 +37,7 @@ const readFileController = async (req, res, next) => {
 
 }
 
-const getAllFilesController = async (req, res,next) => {
+const getAllFilesController = async (req, res, next) => {
     try {
         const files = await File.find({ user: req.userId });
 
@@ -49,13 +49,24 @@ const getAllFilesController = async (req, res,next) => {
 
 const updateFileController = async (req, res, next) => {
     try {
-        const { fileId, content } = req.body;
+        const { fileId } = req.params;
+        const { content } = req.body;
+
+        if (!fileId) return badRequest(res, "File id required");
+
+        if (!content) return badRequest(res, "File content required");
+
+        if (content === undefined) {
+            return badRequest(res, "File content required");
+        }
 
         const file = await File.findOneAndUpdate(
             { _id: fileId, user: req.userId },
             { content },
             { new: true }
         );
+
+        if (!file) return customError(res, 404, "File not found");
 
         return updated(res, file, "File updated");
     } catch (err) {
@@ -68,10 +79,14 @@ const deleteFileController = async (req, res, next) => {
     try {
         const { fileId } = req.params;
 
-        await File.findOneAndDelete({
+        if (!fileId) return badRequest(res, "File id required");
+
+        const file = await File.findOneAndDelete({
             _id: fileId,
             user: req.userId,
         });
+
+        if (!file) return customError(res, 404, "File not found");
 
         return deleted(res, {}, "File deleted");
     } catch (err) {
